Stop star sprite rectangles blocking others out behind them
By disabling depth test when drawing them. The stars are still not depth
sorted for their draw order so now there is a problem where the body of
further away stars may appear in front of closer ones, but that is a
less noticeable effect (only really apparent while the farther star is
lighted up), so will not bother to add full depth-sorting for now.